+++ module/api/front-end/charactermancerStore.js
@@ -8,82 +8,88 @@ import { CalcBaseStats, CalculateStatTotal } from '../../actor/calculations/stat
 export default function({level, tabs, initialTab, species, actor}) {
     const store = new Store({
         actions: {
-            init(context) {
+            async init(context) {
                 const imgSrc = game.settings.get("ptu", "defaultPokemonImageDirectory");
-                if(imgSrc)
-                    GetSpeciesArt(context.state.species, imgSrc).then(imgPath => context.commit('updateArt', imgPath));
-                
+                if(imgSrc) {
+                    const imgPath = await GetSpeciesArt(context.state.species, imgSrc)
+                    await context.commit('updateArt', imgPath);
+                }
+
                 if(context.state.nature) {
                     const natureInfo = game.ptu.natureData[context.state.nature];
                     if(!natureInfo) return;
 
-                    context.commit('updateNature', [context.state.nature, natureInfo]);
+                    await context.commit('updateNature', [context.state.nature, natureInfo]);
                 }
 
-                context.dispatch('changeStats');
+                await context.dispatch('changeStats');
             },
-            changeSpecies(context, species) {
+            async changeSpecies(context, species) {
                 const speciesData = game.ptu.GetSpeciesData(species);
                 if(!speciesData) return;
 
                 const imgSrc = game.settings.get("ptu", "defaultPokemonImageDirectory");
-                if(imgSrc)
-                    GetSpeciesArt(speciesData, imgSrc).then(imgPath => context.commit('updateArt', imgPath));
+                if(imgSrc) {
+                    const imgPath = await GetSpeciesArt(speciesData, imgSrc)
+                    await context.commit('updateArt', imgPath);
+                }
 
-                context.commit('updateSpecies', speciesData);
+                await context.commit('updateSpecies', speciesData);
+                await context.dispatch('changeStats', {speciesData});
             },
-            expChange(context, payload) {
+            async expChange(context, payload) {
                 const exp = Number(payload);
 
-                if(exp === undefined || isNaN(exp)) return;
+                if(exp === undefined || isNaN(exp) || exp == context.state.exp) return;
                 
                 const level = CalcLevel(exp, 50, game.ptu.levelProgression);
+                const oldLevel = duplicate(context.state.level)
 
-                context.commit('updateExp', exp);
-                context.commit('updateLevel', level);
+                await context.commit('updateExp', exp);
+                await context.commit('updateLevel', level);
+                if(oldLevel != level) await context.dispatch('changeStats');
             },
-            levelChange(context, payload) {
+            async levelChange(context, payload) {
                 const level = payload;
-
-                if(level === undefined || isNaN(level)) return;
-
+                if(level === undefined || isNaN(level) || level == context.state.level) return;
+                
                 const exp = game.ptu.levelProgression[level];
 
-                context.commit('updateLevel', level);
-                context.commit('updateExp', exp);
+                await context.commit('updateLevel', level);
+                await context.commit('updateExp', exp);
+                await context.dispatch('changeStats');
             },
-            changeTab(context, payload) {
+            async changeTab(context, payload) {
                 log(`Changing to next tab from: ${payload}`)
                 switch(payload) {
                     case 'species': handleSpeciesPageEnd(context); break;
                     case 'stats': handleStatsPageEnd(context); break;
                 }
             },
-            changeNatureStat(context, {value, isUp}) {
+            async changeNatureStat(context, {value, isUp}) {
                 if(!value) return;
 
                 const otherStat = isUp ? context.state.natureStat.down : context.state.natureStat.up;
                 const natureInfo = Object.entries(game.ptu.natureData).find(x => x[1][0] == (isUp ? value : otherStat) && x[1][1] == (isUp ? otherStat : value))
                 if(!natureInfo) return;
 
-                context.commit('updateNature', natureInfo);
+                await context.commit('updateNature', natureInfo);
             },
-            changeNature(context, nature) {
+            async changeNature(context, nature) {
                 if(!nature) return;
 
                 const natureInfo = game.ptu.natureData[nature];
                 if(!natureInfo) return;
 
-                context.commit('updateNature', [nature, natureInfo])
+                await context.commit('updateNature', [nature, natureInfo])
             },
-            changeStats(context, stats) {
-                
-                const statChanges = mergeObject(context.state.actor.data.data.stats, stats ?? {});
-                console.log(stats, statChanges);
-                const baseStats = CalcBaseStats(statChanges, context.state.species, context.state.actor.data.data.nature.value);
+            async changeStats(context, {levelUpStats, speciesData}={levelUpStats: undefined, speciesData: undefined}) {
+                debug(levelUpStats, speciesData)
+                const statChanges = mergeObject(context.state.actor.data.data.stats, levelUpStats ?? {});
+                const baseStats = CalcBaseStats(statChanges, speciesData ?? context.state.species, context.state.actor.data.data.nature.value);
 
                 // Recalculate stats
-                const levelUpPoints = context.state.actor.data.data.modifiers.statPoints?.total + 10 + context.state.actor.data.data.level.current;
+                const levelUpPoints = context.state.actor.data.data.modifiers.statPoints?.total + 10 + context.state.level;
 
                 const calculatedStats = CalculateStatTotal(levelUpPoints, baseStats, {ignoreStages: true});
                 const result = {
@@ -91,31 +97,31 @@ export default function({level, tabs, initialTab, species, actor}) {
                     levelUpPoints: calculatedStats.levelUpPoints
                 };
 
-                context.commit('updateStats', result);
+                await context.commit('updateStats', result);
             }
         },
         mutations: {
-            updateLevel(state, level) {
+            async updateLevel(state, level) {
                 state.level = level;
                 return state;
             },
-            updateExp(state, exp) {
+            async updateExp(state, exp) {
                 state.exp = exp;
                 return state;
             },
-            updateSpecies(state, species) {
+            async updateSpecies(state, species) {
                 state.species = species;
                 return state;
             },
-            updateArt(state, imgPath) {
+            async updateArt(state, imgPath) {
                 state.imgPath = imgPath;
                 return state;
             },
-            updateTab(state, tab) {
+            async updateTab(state, tab) {
                 state.currentTab = tab;
                 return state;
             },
-            updateNature(state, natureInfo) {
+            async updateNature(state, natureInfo) {
                 state.nature = natureInfo[0];
                 state.natureStat = {
                     up: natureInfo[1][0], 
@@ -123,7 +129,7 @@ export default function({level, tabs, initialTab, species, actor}) {
                 };
                 return state;
             },
-            updateStats(state, {stats, levelUpPoints}) {
+            async updateStats(state, {stats, levelUpPoints}) {
                 state.stats = stats;
                 state.levelUpPoints = levelUpPoints;
                 return state;
@@ -162,8 +168,8 @@ async function handleSpeciesPageEnd(context) {
         await Dialog.confirm({
             title: `Evolution Detected!`,
             content: `<p class='readable pb-2 pt-1'>Wow! It looks like ${context.state.actor.name} is about to evolve into<br><br><b>${species._id}</b>!<br><br>Will you let it?</p>`,
-            yes: _ => {
-                context.dispatch('changeSpecies', species._id)
+            yes: async _ => {
+                await context.dispatch('changeSpecies', species._id)
             },
             rejectClose: false
         });
@@ -174,7 +180,7 @@ async function handleSpeciesPageEnd(context) {
     }
 
     context.state.tabs.activate('stats');
-    context.commit('updateTab', 'stats');
+    await context.commit('updateTab', 'stats');
 }
 
 async function handleStatsPageEnd() {